LPD-49732 Create base class for common code
