+++ modules/apps/site/site-cms-site-initializer/bnd.bnd
@@ -2,3 +2,4 @@ Bundle-Name: Liferay Site CMS Site Initializer
 Bundle-SymbolicName: com.liferay.site.cms.site.initializer
 Bundle-Version: 1.0.5
 Web-ContextPath: /site-cms-site-initializer
+-dsannotations-options: inherit
