+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/CategorizationSectionFragmentRenderer.java
@@ -8,13 +8,7 @@ package com.liferay.site.cms.site.initializer.internal.fragment.renderer;
 import com.liferay.fragment.renderer.FragmentRenderer;
 import com.liferay.fragment.renderer.FragmentRendererContext;
 import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
-import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.GroupConstants;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.theme.ThemeDisplay;
-import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.site.cms.site.initializer.internal.configuration.CMSSiteInitializerConfiguration;
 import com.liferay.site.cms.site.initializer.internal.display.context.CategorizationSectionDisplayContext;
 
@@ -22,7 +16,6 @@ import java.io.IOException;
 
 import java.util.Locale;
 import java.util.Map;
-import java.util.Objects;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletContext;
@@ -41,7 +34,8 @@ import org.osgi.service.component.annotations.Reference;
 	configurationPid = "com.liferay.site.cms.site.initializer.internal.configuration.CMSSiteInitializerConfiguration",
 	service = FragmentRenderer.class
 )
-public class CategorizationSectionFragmentRenderer implements FragmentRenderer {
+public class CategorizationSectionFragmentRenderer
+	extends BaseSectionFragmentRenderer {
 
 	@Override
 	public String getCollectionKey() {
@@ -51,30 +45,6 @@ public class CategorizationSectionFragmentRenderer implements FragmentRenderer {
 	@Override
 	public String getLabel(Locale locale) {
 		return _language.get(locale, "categorization");
-	}
-
-	@Override
-	public boolean isSelectable(HttpServletRequest httpServletRequest) {
-		ThemeDisplay themeDisplay =
-			(ThemeDisplay)httpServletRequest.getAttribute(
-				WebKeys.THEME_DISPLAY);
-
-		if (!FeatureFlagManagerUtil.isEnabled(
-				themeDisplay.getCompanyId(), "LPD-17564")) {
-
-			return false;
-		}
-
-		Group group = _groupLocalService.fetchGroup(
-			themeDisplay.getScopeGroupId());
-
-		if ((group == null) ||
-			!Objects.equals(group.getGroupKey(), GroupConstants.CMS)) {
-
-			return false;
-		}
-
-		return true;
 	}
 
 	@Override
@@ -110,9 +80,6 @@ public class CategorizationSectionFragmentRenderer implements FragmentRenderer {
 
 	private volatile CMSSiteInitializerConfiguration
 		_cmsSiteInitializerConfiguration;
-
-	@Reference
-	private GroupLocalService _groupLocalService;
 
 	@Reference
 	private Language _language;
